pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "garmin-cron-worker"
version = "0.1.0"
description = "Cron worker that loads tasks and fetches Garmin Connect metrics."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "garminconnect",
  "garth",
  "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

